Revert "Refactoring and using BD"
