--- src/bot/index.ts
@@ -1,3 +1,0 @@
-import Bot from './bot';
-
-export default Bot;
