+++ src/model.ts
@@ -2,6 +2,7 @@ import fs from 'fs';
 import { Color, NumberFormat } from 'google-spreadsheet';
 import path from 'path';
 import { MODEL_PATH } from './constants';
+import { armeniaSheetText, russiaSheetText } from './textUtils';
 
 const ENCODING_TYPE = 'utf8';
 
@@ -19,11 +20,6 @@ export interface IModel {
     nameColumnName: ICategory;
     modelSheetName: string;
     rowColors: IRowColors;
-    documents: Array<{
-        id: string;
-        text: string;
-        currency: string;
-    }>
 }
 
 export interface IRowColors {
@@ -63,7 +59,15 @@ export default class Model {
     }
 
     get documents() {
-        return this.JSONModel.documents;
+        return [{
+            text: armeniaSheetText,
+            currency: '֏',
+            id: process.env.ARMENIA_SPREADSHEET_ID,
+        }, {
+            text: russiaSheetText,
+            currency: '₽',
+            id: process.env.MAIN_SPREADSHEET_ID,
+        }];
     }
 
     private get modelFilePath() {
