+++ src/types/environment.d.ts
@@ -2,6 +2,8 @@ declare global {
     namespace NodeJS {
         interface ProcessEnv {
             BOT_TOKEN: string;
+            ARMENIA_SPREADSHEET_ID: string;
+            MAIN_SPREADSHEET_ID: string;
             PRIVATE_KEY: string;
             CLIENT_EMAIL: string;
             URL: string;
